+++ CHANGES.rst
@@ -1,4 +1,21 @@
 Changes
 
-Unreleased
-----------
+0.1.15
+------
+- add hiv and indicators followup forms
+- add additional items to lists, order
+
+0.1.14
+------
+- update arv regimen list data based on "other" values entered
+- change m2ms to blank=True and use form logic `required_if`
+- fix incorrect base class on some admin classes
+- other minor form logic fixes
+
+0.1.13
+------
+- second live version
+
+0.1.8
+-----
+- first live version